+++ R/utils_model_function.R
@@ -19,6 +19,7 @@ get_model_function <- function(model) {
     inherits(model, "lme") ~ "lm",
     inherits(model, "glmerMod") ~ "glm",
     inherits(model, "nlmerMod") ~ "lm",
+    inherits(model, c("zeroinfl", "hurdle")) ~ "glm",
     inherits(model, c("lmerMod", "merModLmerTest")) ~ "lm",
     TRUE ~ "glm"
   )
@@ -46,6 +47,7 @@ get_predict_function <- function(model) {
     inherits(model, "vglm") ~ "vglm",
     inherits(model, "glm") ~ "glm",
     inherits(model, "lm") ~ "lm",
+    inherits(model, c("zeroinfl", "hurdle")) ~ "zeroinfl",
     TRUE ~ "generic"
   )
 }